client/client/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
clipboard==0.0.4
commonmark==0.9.1
comtypes==1.1.14
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
idna==3.4
Levenshtein==0.20.9
openpyxl==3.0.10
pydantic==1.10.4
Pygments==2.14.0
pyperclip==1.8.2
python-dotenv==0.21.0
python-Levenshtein==0.20.9
# uncomment below two on Windows for pdf filler
# pywin32==305
# pywinauto==0.6.8
rapidfuzz==2.13.7
requests==2.28.2
rich==13.1.0
six==1.16.0
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.14
XlsxWriter==3.0.7
beautifulsoup4==4.11.1
prompt_toolkit==3.0.36
schedule==1.1.0
aiohttp==3.8.4
pymongo==4.3.3
email-validator==1.3.1
openai==0.27.2
plotly==5.13.1
pandas==1.5.3